Start `kind` to provide a full kubernetes cluster
Users need to provide the `kind` binary at
`kind-release/kind-linux-amd64`. This can be done by e.g. using a
`github-release` resource as the `kind-release` `get` step. However, any
other mechanism to provide that one binary at that path will also do.

Essentially, this binary will be used to run `kind create cluster`. By
default, `kind` will spin up a cluster with the default node image which
gets pulled of docker hub (provided by the kind team).
`kubectl` will be downloaded in the latest stable version from
kubernetes upstream and made available in the `$PATH`.

However, users can also provide a kubernetes source tree which this task
will then build the node image off of. This can be done by using a
`git` resource as the `kubernetes-git` `get` step.
This will compile all the needed kubernetes components and build the
node image based on that. Also `kubectl` will be compiled in the same
version and made available in the `$PATH`.

When `kind` has been started, `$KUBECONFIG` will be exported, so that
all clients from there on should have acces to the correct coordinates
to the APIServer.

To support running node images with newer versions of kubernetes
(1.15+) we need to apply a little hack to the node image after it has
been created by `kind`:
The `kubelet` wont start unless `/dev/kmsg` is available, this seems to
be due to a change/removal(?) of `cadvisor`. Therefore we inject a
change into the `kubelet` `systemd unit` to do the following before
starting the `kubelet`: Check, if that file is available -- if not, try
to create a link to `/dev/console`.