Make task reusable
... by making the user's test configuarble from the outside,
specifiaclly by specifying them in `$KIND_TESTS`. This will be called as
`bash -c "$KIND_TESTS"`, so this can either point to a different
executable or hold any any comands bash can handle.

If that variable is not setup, we bail out early and don't even attempt
to start docker or kind.