Use a special flannel deployment config
We are not pulling the flannel deployment config from upstream
("github") but use a local one, which uses apps/v1 DaemonSet and adds
some configuration ('spec.selector') which the new API demands.

With that flannel should work on current master, where the DaemonSet
v1beta1 seems to have been removed.