Fix /dev/kmsg linking regardless of which image is used
The previous approach to link `/dev/kmsg` to `/dev/console` only works on
images that we build ourselves.

The new approach is to do the linking at runtime, parallel to kind
brining up the cluster:
We run a function parallel to kind, which tries to exec into the control
plane node and do the linking. This should succeed as soon as the node
container is up, in which case the function will terminate gracefully.

As this is fixed in kind already [0], we can remove that hack probably
when kind 0.5.0 is out.

Why is this linking even needed? See also [0]!

[0]: https://github.com/kubernetes-sigs/kind/issues/662