Reafctor of pipeline
The building of the image is now a separate flow. The image will only be
updated when the Dockerfile is updated or when there is a new alpine
image available.

The task config will now be updated with the specific tag of the latest
image build. From now on, you can check out any commit of master and it
will use a specific version of the task image instead of latest.

The pipeline configuration has been cleaned up and simplified, most
notably by:
- saner usage of yaml anchors
- renaming task and jobs to better reflect what they do