Make the task output kind-{image,rootfs}
Those outputs should provide the node image as a rootfs (e.g. to be
consumed by another task) and the OCI image (e.g. to be uploaded via the
registry-image-resource).

The rootfs should only be exported if `KIND_EXPORT_ROOTFS` is set, the
OCI image should only be exported it `KIND_EXPORT_IMAGE` is set.

The image or the rootfs will be placed into a subdirectory `node`. This
allows us to potentially export more/different images (potentially a
loadbalancer image, potentially a CNI image, ...). On the flip side, you
cannot directly consume `kind-rootfs` but need to use another task which
only ouputs the node image. We will see what approach is better ...