Use 'node-image' input if configured
If `$KIND_TESTS` is empty, we don't attempt to start kind and will still
return a successful exit code.

We do this, so you can have a task which just builds (and exports) a
node image but does not actually start a cluster. The exported node
image can be used later in other tasks.