Don't spam output with colors on log::init
When xtrace is set and the shared library gets loaded and logging
enabled (`. lib.inc.sh && log::init`), the output gets spammed with
colored output. We want to avoid that to not confues users and disable
xtrace in the `log::init` but reset xtrace at the end of that function
to its previous state. Therefore everything will still run with xtrace
if enabled, only for the `log::init` xtrace will be forced to be
disabled.