[ci skip] Check resources when setting the pipeline
Example:
After a new kubernetes version has been released, we change the pipeline
to use a different branch for the resources `k8s-git-current{,-minus-1,minus-2}`.
Because those resource are not checked that often,we now might end up
with with a resource version (git revision) that is not on the
resource's branch after we changed that.
As a workaround we ask concourse to check those resources right after we
set the pipeline, which should give us a valid resource version.