Introduce a node prepper instead of the kmsg linker
For installations where concourse runs with btrfs the kubelet has issues
with coming up, as it expects to interact with `/dev/loop0`. This device
might not exist.

Therefore we externalised the former kmsg linker script and made it more
generic. It now does not only do the linking of `/dev/kmsg`, but also
creates `/dev/loop0` if it is not present.
Users are also able to extend the node-prepper script by appending to it
in `KIND_PRE_START`.

But most notable: I am not sure why the kubelet needs to interact with
`/dev/loop0` and what the implications are by just creating that device.
I could confirm that the kubelet and thus the cluster comes up with that
hack, but I am not sure which other issues this might have opened up
now. I might try to poke around in k/k to get a better understanding why
the kubelet expect this device, what it does with it and why it bails if
it's not there.