Let concourse download `oq`
We can use the github-release resource type and use the github token, we
already use elsewhere in the pipeline.

This means, one less thing that is not managed by concourse.

But it also means, that when people build custom base images somewhat
based on our Dockerfile, they now need to change their tooling to also
download the binary up front and put it into the build context's
`./external/oq/` directory.