Less surprises
We don't use a temporary kube config anymore, we just let the tools run
with the default.

Also, we append our bin directory, where we place the kubectl and kind
binary into, to the PATH and make that permanent in the `~/.bashrc`.
We even set an alias there, to make kubectl available as `k`.

This all should make debugging, e.g. when you `fly intercept` the task
container, a bit nicer and removes some surprises (where's the kube
config?).